Implement SecondVote into ElectionChances
