+++ src/components/ElectionChancesCard.css
@@ -16,10 +16,6 @@
 	padding-right: 2%;
 }
 
-
-#img {
-	display: none;
-}
 .election-chances-card-subtitle {
 	font-size: 14px;
 	font-family: 'protipo', sans-serif;
