+++ src/pages/ElectionChances.css
@@ -101,7 +101,7 @@
 	font-size: 130%;
 }
 
-.grey-back {
+.black-back {
 	background-color: #181924;
 }
