+++ src/pages/ElectionChances.tsx
@@ -1,10 +1,10 @@
-import { IonButtons, IonContent, IonPage, IonTabs, IonTabBar, IonTabButton, IonIcon, IonTitle, IonBackButton, IonText, IonToolbar, IonSegment, IonSegmentButton, IonCardSubtitle, IonCardTitle } from '@ionic/react';
-import React from 'react';
+import { IonButtons, IonContent, IonPage, IonTabs, IonTabBar, IonTabButton, IonIcon, IonTitle, IonBackButton, IonText, IonToolbar, IonSegment, IonSegmentButton, IonCardSubtitle, IonCardTitle, useIonViewDidEnter } from '@ionic/react';
+import React, { useState } from 'react';
 import ElectionchancesCard from '../components/ElectionChancesCard';
-import StateListCard from '../components/StateListCard';
+import SecondVoteCard from '../components/SecondVoteCard';
 import SegmentButtons from '../components/SegmentButtons';
 import './ElectionChances.css';
-import { Candidate } from '../Types';
+import { Candidate, StateList } from '../Types';
 
 interface ElectionchancesProps {
 	candidate: Candidate;
@@ -20,6 +20,80 @@ const Electionchances: React.FC<ElectionchancesProps> = ({ candidate }: Election
 	}
 
 	const [tab, setTab] = React.useState('0'); // eslint-disable-line @typescript-eslint/no-unused-vars
+/* Second Vote Code */
+
+const [fixedPosition, setFixedPosition] = useState<number | undefined>();
+	const [previousScrollHeight, setpreviousScrollHeight] = useState<number>(0);
+	const [lastCandidate, setLastcandidate] = useState<number>(0);
+
+	const initialFixedPositionHandler = (): undefined | number => {
+		const windowHeight: number | undefined = document.getElementById('second-vote-results')
+			?.clientHeight;
+		const cardHeight: number | undefined = document.getElementById('second-vote-card')
+			?.clientHeight;
+		console.log(cardHeight, windowHeight);
+		if (typeof windowHeight === 'number' && typeof cardHeight === 'number') {
+			const halfWindowHeight = windowHeight / 2;
+			const cardNums = Math.round(halfWindowHeight / cardHeight);
+			return cardNums + 1;
+		}
+	};
+
+	useIonViewDidEnter(() =>
+		setTimeout(function () {
+			setFixedPosition(initialFixedPositionHandler());
+			setLastCandidateHandler();
+		}, 100)
+	);
+
+	const cardNumsHandler = (voteRes: { name: string; electionChance: number }[]): number => {
+		const cardNums = voteRes.length;
+		return cardNums;
+	};
+
+	const setFixedPositionHandler = (): void => {
+		const cardNums = cardNumsHandler(candidate.secondVote);
+		const voteResults = document.getElementById('second-vote-results');
+		const initialPos = initialFixedPositionHandler();
+		const cardHeight = 57;
+		const cardblank = 10;
+		const oneScroll = cardHeight + cardblank;
+		if (!fixedPosition || !voteResults || !initialPos) {
+			return;
+		}
+		//margin top 10px
+		const scrollHeight = voteResults.scrollTop - 10;
+		const scrollDistance = Math.abs(scrollHeight - previousScrollHeight);
+
+		let direction;
+		if (previousScrollHeight < scrollHeight) {
+			direction = 1;
+		} else {
+			direction = -1;
+		}
+		//If a user scroll at the bottom of screen
+		if (fixedPosition === cardNums && scrollDistance >= oneScroll) {
+			setFixedPosition(fixedPosition - 1);
+		}
+
+		if (fixedPosition < cardNums && scrollDistance >= oneScroll) {
+			setpreviousScrollHeight(scrollHeight);
+			const scrollNums = (scrollDistance / oneScroll) >> 0;
+			setFixedPosition(fixedPosition + scrollNums * direction);
+		}
+		//If a user try to scroll up above the initial position
+		if (fixedPosition < initialPos) {
+			setFixedPosition(fixedPosition + 1);
+		}
+	};
+
+	const setLastCandidateHandler = (): void => {
+		//if we get deta from backend, we should write calculation logic here
+		//But now I detemine last candidate is rank #4
+		//this function should excute after rendering this page
+		setLastcandidate(4);
+	};
+
 
 	/* This is returned when using this component */
 	return (
@@ -44,9 +118,9 @@ const Electionchances: React.FC<ElectionchancesProps> = ({ candidate }: Election
 					</div>
 					} */}
 						<IonContent>
-			<div className="grey-back">		
+			<div className="black-back">		
            
-					{firstTab ?
+					{firstTab=="0" ?
 					<ElectionchancesCard firstCandidate={true} vote={candidate.electionResults[0]} />
 					: null
 					}
